fixed issue with single day ranges not displaying.
thanks to david for pointing this out to me.
if not specified, hh:mm is assumed to be 00:00, so at 00:01, we're out of the date range and the page will be incorrect. as this is, the page will still be incorrect between 23:59 and midnight for the same reason. thought about looking up sunrise and sunset times for single days, but ain't nobody got time fo' dat.